fixed bugs add cartItems

+++ Interface/ICartRepository.cs
@@ -6,7 +6,7 @@ public interface ICartRepository
 {
     Task<Cart?> GetCartByUserIdAsync(Guid UserId);
     Task<CartItem?> AddCartItemAsync(Guid UserId,Guid ProductId,int quantity);
-    Task<CartItem?> UpdateCartItemAsync(Guid CartItemId, decimal quantity);
+    Task<CartItem?> UpdateCartItemAsync(Guid CartItemId, int quantity);
     Task<bool> RemoveCartItemAsync(Guid CartItemId);
     Task<bool> ClearCartAsync(Guid userId);
